more int overflow risks (#1101)
* more int overflow risks

* some reverts due to test issues

* try to fix skip issues

* Update poi-integration-exceptions.csv

+++ poi-ooxml/src/main/java/org/apache/poi/xwpf/usermodel/XWPFRun.java
@@ -807,7 +807,7 @@ public class XWPFRun implements ISDTContents, IRunElement, CharacterRun {
         if (pr == null || pr.sizeOfKernArray() == 0) {
             return 0;
         }
-        return (int)POIXMLUnits.parseLength(pr.getKernArray(0).xgetVal());
+        return Math.toIntExact(POIXMLUnits.parseLength(pr.getKernArray(0).xgetVal()));
     }
 
     @Override
